refactor(techdocs): group components files
