+++ plugins/techdocs/src/reader/components/TechDocsReaderPageHeader/TechDocsReaderPageHeader.test.tsx
@@ -18,7 +18,7 @@ import { TechDocsReaderPageHeader } from './TechDocsReaderPageHeader';
 import { act } from '@testing-library/react';
 import { renderInTestApp } from '@backstage/test-utils';
 import { entityRouteRef } from '@backstage/plugin-catalog-react';
-import { rootRouteRef } from '../../routes';
+import { rootRouteRef } from '../../../routes';
 
 describe('<TechDocsReaderPageHeader />', () => {
   it('should render a techdocs page header', async () => {
