+++ plugins/techdocs/src/reader/components/TechDocsReaderPageHeader/TechDocsReaderPageHeader.tsx
@@ -26,8 +26,8 @@ import {
   getEntityRelations,
 } from '@backstage/plugin-catalog-react';
 
-import { rootRouteRef } from '../../routes';
-import { TechDocsEntityMetadata, TechDocsMetadata } from '../../types';
+import { rootRouteRef } from '../../../routes';
+import { TechDocsEntityMetadata, TechDocsMetadata } from '../../../types';
 
 /**
  * Props for {@link TechDocsReaderPageHeader}
